Ex 5 and 6 completely done

+++ lab4/src/sum.c
@@ -1,7 +1,7 @@
 #include "sum.h"
 
-int Sum(const struct SumArgs *args) {
-  int sum = 0;
+intmax_t Sum(const struct SumArgs *args) {
+  intmax_t sum = 0;
   // TODO: your code here 
 
   for(int i = args->begin;i < args->end;i++)
@@ -14,5 +14,5 @@ int Sum(const struct SumArgs *args) {
 
 void *ThreadSum(void *args) {
   struct SumArgs *sum_args = (struct SumArgs *)args;
-  return (void *)(unsigned int)Sum(sum_args);
+  return (void *)(intmax_t)Sum(sum_args);
 }
