+++ lab4/src/sum.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #ifndef THREAD_SUM_H
 #define THREAD_SUM_H
 
@@ -8,7 +10,7 @@ struct SumArgs
   int end;
 };
 
-int Sum(const struct SumArgs *args);
+intmax_t Sum(const struct SumArgs *args);
 void *ThreadSum(void *args);
 
 #endif
